refactor: split generator core operations
